+++ tools/build-package.sh
@@ -82,6 +82,7 @@ find $source_roots -type f \
   ! -path '*/.build-m14/*' \
   ! -path '*/.build-m15/*' \
   ! -path '*/.build-m16/*' \
+  ! -path '*/.build-m17/*' \
   ! -path '*/.libs/*' \
   ! -path '*/.deps/*' \
   -print | LC_ALL=C sort > "$manifest"